Fixed Input Fields
Changed the system to using List's and instead of opening a window that had every question. I made it so that they can open a window to each individual question. This got rid of the need to use a for loop because that doesn't work with gui fields, and it makes it easier to see every question as they are viewed individually now. So now they can actually continue to add in questions as they like one by one.